+++ library/ssl_cache.c
@@ -40,7 +40,7 @@
 #endif
 
 #include "mbedtls/ssl_cache.h"
-#include "mbedtls/ssl_internal.h"
+#include "ssl_internal.h"
 
 #include <string.h>
