+++ library/ssl_cli.c
@@ -36,7 +36,7 @@
 #endif
 
 #include "mbedtls/ssl.h"
-#include "mbedtls/ssl_internal.h"
+#include "ssl_internal.h"
 #include "mbedtls/debug.h"
 #include "mbedtls/error.h"
