+++ library/ssl_cookie.c
@@ -39,7 +39,7 @@
 #endif
 
 #include "mbedtls/ssl_cookie.h"
-#include "mbedtls/ssl_internal.h"
+#include "ssl_internal.h"
 #include "mbedtls/error.h"
 #include "mbedtls/platform_util.h"
